saifur 32(database)
member_db package

+++ packages_list.php
@@ -1,7 +1,7 @@
 <?php
 session_start(); // this NEEDS TO BE AT THE TOP of the page before any output etc
 $showname = $_SESSION['uname'];
-
+$designation = $_SESSION['profation'];
 if ($_GET != NULL && ($_GET['un'] != 'u' && $_GET['un'] != 'i' && $_GET['un'] != 'd' && $_GET['un'] != 'w')) {
   $uname = $_GET['un'];
 } else {
@@ -38,13 +38,38 @@ if (!$conn) {
       header("Location: packages_list.php?un=i");
     }
     if (isset($_POST['pkg_id'])) {
-      $pkg_id = $_POST['pkg_id'];
-      $sql = "DELETE FROM package WHERE pkg_id = '$pkg_id'";
-      $stid = oci_parse($conn, $sql);
-      $r = oci_execute($stid);
+      if($designation == 'Member') {
+        $pkg_id = $_POST['pkg_id'];
+        $sql = "select * from users where USERNAME = '$uname'";
+        $stid = oci_parse($conn, $sql);
+        $r = oci_execute($stid);
+        $row = oci_fetch_array($stid, OCI_ASSOC + OCI_RETURN_NULLS);
+        $br_name = $row["BR_NAME"];
+        $_SESSION['xxx'] = $br_name;
+        $sql = "select *from income order by trx_id desc";
+        $stid = oci_parse($conn, $sql);
+        $r = oci_execute($stid);
+        $row = oci_fetch_array($stid, OCI_ASSOC + OCI_RETURN_NULLS);
+        $trx_id = $row['TRX_ID'] + 1;
+        $sql = "select * from package where pkg_id = '$pkg_id'";
+        $stid = oci_parse($conn, $sql);
+        $r = oci_execute($stid);
+        $row = oci_fetch_array($stid, OCI_ASSOC + OCI_RETURN_NULLS);
+        $amount = $row['PKG_CHARGE'];
+        $sql = "insert into income (trx_id, username, inc_amount, br_name, inc_type, inc_dateandtime) values($trx_id, '$uname', $amount, '$br_name', 'Member Payment', SYSTIMESTAMP)";
+        $stid = oci_parse($conn, $sql);
+        $r = oci_execute($stid);
+      }
+      else {
+        $pkg_id = $_POST['pkg_id'];
+        $sql = "DELETE FROM package WHERE pkg_id = '$pkg_id'";
+        $stid = oci_parse($conn, $sql);
+        $r = oci_execute($stid);
+        header("Location: packages_list.php?un=d");
+      }
       // $x = $pkg_id;
 
-      header("Location: packages_list.php?un=d");
+      
     }
     if (isset($_POST['pkg_id2'])) {
       $pkg_id = $_POST['pkg_id2'];
@@ -103,6 +128,110 @@ if (!$conn) {
     </nav>
     <!-- /.navbar -->
     <?php
+    if($designation == 'Member') {
+      echo '
+      <!-- Main Sidebar Container -->
+      <aside class="main-sidebar sidebar-dark-primary elevation-4">
+        <!-- Brand Logo -->
+        <a href="#" class="brand-link">
+          <img src="dist/img/AdminLTELogo.png" alt="AdminLTE Logo" class="brand-image img-circle elevation-3" style="opacity: .8">
+          <span class="brand-text font-weight-light">Fitness Mania</span>
+        </a>
+  
+        <!-- Sidebar -->
+        <div class="sidebar">
+          <!-- Sidebar user panel (optional) -->
+          <div class="user-panel mt-3 pb-3 mb-3 d-flex">
+            <div class="image">
+              <img src="dist/img/user2-160x160.jpg" class="img-circle elevation-2" alt="User Image">
+            </div>
+            <div class="info">
+              <a href="member_profile.php" class="d-block">'.$uname.'
+              </a>
+            </div>
+          </div>
+  
+          <!-- SidebarSearch Form -->
+          <!-- <div class="form-inline">
+            <div class="input-group" data-widget="sidebar-search">
+              <input class="form-control form-control-sidebar" type="search" placeholder="Search" aria-label="Search">
+              <div class="input-group-append">
+                <button class="btn btn-sidebar">
+                  <i class="fas fa-search fa-fw"></i>
+                </button>
+              </div>
+            </div>
+          </div> -->
+  
+          <!-- Sidebar Menu -->
+          <nav class="mt-2">
+            <ul class="nav nav-pills nav-sidebar flex-column" data-widget="treeview" role="menu" data-accordion="false">
+              <!-- Add icons to the links using the .nav-icon class
+                 with font-awesome or any other icon font library -->
+              <li class="nav-item">
+                <a href="#" class="nav-link">
+                  <i class="nav-icon fas fa-tachometer-alt"></i>
+                  <p>
+                    Dashboard
+                    <i class="right fas fa-angle-left"></i>
+                  </p>
+                </a>
+                <ul class="nav nav-treeview">
+  
+                  <li class="nav-item">
+                    <a href="member_db.php" class="nav-link">
+                      <i class="far fa-circle nav-icon"></i>
+                      <p>Member</p>
+                    </a>
+                  </li>
+  
+                </ul>
+              </li>
+  
+  
+  
+  
+  
+              <li class="nav-item">
+                <a href="#" class="nav-link">
+                  <i class="nav-icon far fa-envelope"></i>
+                  <p>
+                    Mailbox
+                    <i class="fas fa-angle-left right"></i>
+                  </p>
+                </a>
+                <ul class="nav nav-treeview">
+                  <li class="nav-item">
+                    <a href="pages/mailbox/mailbox.html" class="nav-link">
+                      <i class="far fa-circle nav-icon"></i>
+                      <p>Inbox</p>
+                    </a>
+                  </li>
+                  <li class="nav-item">
+                    <a href="pages/mailbox/compose.html" class="nav-link">
+                      <i class="far fa-circle nav-icon"></i>
+                      <p>Compose</p>
+                    </a>
+                  </li>
+  
+                </ul>
+              </li>
+              
+  
+  
+  
+            </ul>
+            </li>
+  
+            </ul>
+          </nav>
+          <!-- /.sidebar-menu -->
+        </div>
+        <!-- /.sidebar -->
+      </aside>  
+      ';
+    }
+    else {
     if ($_GET == NULL || ($_GET != NULL && ($_GET['un'] == 'd' || $_GET['un'] == 'w' || $_GET['un'] == 'i' || $_GET['un'] == 'u'))) {
       echo '
         <!-- Main Sidebar Container -->
@@ -335,6 +464,7 @@ if (!$conn) {
           
   ';
     }
+  }
     ?>
     <!-- Content Wrapper. Contains page content -->
     <div class="content-wrapper">
@@ -395,12 +525,14 @@ if (!$conn) {
             <div class="col-lg-6 col-md-12" style="padding-top: 15px;padding-right:40px;">
               <!-- Insert Modal -->
               <?php
+              if($designation <> 'Member') {
               if ($_GET == NULL || ($_GET != NULL && ($_GET['un'] == 'd' || $_GET['un'] == 'w' || $_GET['un'] == 'i' || $_GET['un'] == 'u'))) {
                 echo '
                 <button type="button" class="insert btn btn-success float-right" data-toggle="modal" data-target="#exampleModal">Add New</button>
               
                 ';
               }
+            }
               ?>
               <!-- Modal -->
               <div class="modal fade" id="exampleModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
@@ -545,12 +677,17 @@ if (!$conn) {
               <td>" . $row["PKG_TYPE"] . "</td>
               <td>" . $row["PKG_CHARGE"] . "</td>
               <td>" . $row["PKG_DURATION"] . "</td>";
+              if($designation == 'Member') {
+                echo "<td> <button class='delete btn btn-sm btn-success'>Purchase</button>";
+              }
+              else {
                 if ($_GET == NULL || ($_GET != NULL && ($_GET['un'] == 'd' || $_GET['un'] == 'w' || $_GET['un'] == 'i' || $_GET['un'] == 'u'))) {
                   echo "<td> <button class='delete btn btn-sm btn-danger'>Remove</button> <button class='update btn btn-sm btn-primary' id=" . $row['BR_NAME'] . ">Edit</button></td>";
                 }
                 echo "</tr>
               ";
               }
+            }
               ?>
             </tbody>
           </table>
